server/pyproject.toml
[project]
name = "lumina"
version = "0.1.0"
description = "Lumina Logic Server"
authors = [{ name = "Lumina Team" }]
dependencies = [
    "websockets>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "psutil>=6.0.0",
    "aiofiles>=24.0.0",
    "mss>=9.0.1",
    "easyocr>=1.7.1",
    "pywin32>=306",
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "uiautomation>=2.0.15",
    "numpy<2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lumina*"]

[project.scripts]
lumina = "lumina.__main__:main"
